+++ webpack.prod.local.js
@@ -2,6 +2,7 @@ let config = require('./webpack.config.js');
 var path = require('path');
 const CleanWebpackPlugin = require('clean-webpack-plugin');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
+const ManifestPlugin = require('webpack-manifest-plugin');
 
 config.output = {
     filename: '[name].[contenthash].js',
@@ -20,7 +21,25 @@ config.plugins.push(
         template: './packages/server/template.html',
         minify: true,
         cache: true
-    })
+    }),
+    new ManifestPlugin({
+        fileName: 'manifest.json',
+        basePath: process.env.NODE_ENV === 'production' ? '/build/' : '/public/',
+        seed: {
+          name: 'App name',
+          short_name: 'Short app name',
+          icons: '',
+          start_url: '',
+          scope: '.',
+          display: 'standalone',
+          orientation: 'portrait-primary',
+          background_color: '#fff',
+          theme_color: '#3f51b5',
+          description: 'description app',
+          dir: 'ltr',
+          lang: 'en-US'
+        }
+    })    
 )
 
 config.optimization = Object.assign(config.optimization, {