+++ webpack.prod.sw.js
@@ -45,7 +45,7 @@ config.plugins.push(
           dir: 'ltr',
           lang: 'en-US'
         }
-    })    
+    })
 )
 
 config.optimization = Object.assign(config.optimization, {